Web conf for loca nginx improvement

+++ tools_generator.sh
@@ -245,7 +245,9 @@ function apply_nginx_user_defined_values(){
           #sed -i.bak "s/$ENVIRONMENT_DOCKER_IMAGE_VERSION/$ENVIRONMENT_DOCKER_IMAGE_VERSION_OVERRIDE/" $CONFIGMAP_FILE_PATH
 
     sed -i.bak "s/server_name.*\#Server.*/server_name $HEX_CONFIGMAP_API_HOST; \#Server domain/" $TEMPLATE_GENERATE_PATH/local/nginx/nginx.conf
-    rm $TEMPLATE_GENERATE_PATH/local/nginx/conf.d/web.conf.bak
+    if [[ -f "$TEMPLATE_GENERATE_PATH/local/nginx/conf.d/web.conf.bak" ]]; then 
+      rm $TEMPLATE_GENERATE_PATH/local/nginx/conf.d/web.conf.bak
+    fi
 
     CLIENT_DOMAIN=$(echo $HEX_CONFIGMAP_DOMAIN | cut -f3 -d "/")
     sed -i.bak "s/server_name.*\#Client.*/server_name $CLIENT_DOMAIN; \#Client domain/" $TEMPLATE_GENERATE_PATH/local/nginx/conf.d/web.conf
